requirements.txt
paho-mqtt
rhasspy-hermes
boto3